Add search functionality for owners and pets by name
I have created the "SearchPage" component in React to allow users to search for owners and pets by name.
I have used the React "useState" hook to manage the search state and the "axios" library to make HTTP requests.
The "handleSearch" function is called when the user clicks the "Search" button and makes two API requests to get the results filtered by the search term.
I have displayed the search results in a list with information about owners and pets.