Format admission date in search results
Update the SearchPage component to format the admission date for pets in the "dd/mm/yyyy" format. This was achieved by adding a call to the `toLocaleDateString` function with the `'es-AR'` parameter to create a new date object and format it in the desired format. The change was made to the following line of code:

<p>Fecha ingreso: {new Date(pet.admission).toLocaleDateString('es-AR')}</p>

This improves the readability of the search results and makes the date format consistent with regional settings in Argentina