Changes to the pets.js API
A new field "observations" was added to the "pets" table.
Two fields, "observations" and "discharge", were added to the POST and PUT method to allow for tracking of pet details after being discharged.
The "instructions" field is now also included in the POST and PUT method.
Two fields, "observations" and "discharge", were added to the PUT method to enable updating of pet details after being discharged.
Overall, these changes allow for better tracking of the health and well-being of pets in the system, and facilitate their tracking in case of future updates.