A new API called appointments.js was created.
The API is capable of handling the following HTTP requests: GET, POST, PUT, and DELETE. If the request is GET, the API returns a list of appointments (or an individual appointment if an ID is provided). If the request is POST, a new appointment is created. If the request is PUT, an existing appointment is updated. If the request is DELETE, an existing appointment is deleted.

The API also includes data relationships for Owner and Pet, and uses the include notation to include this data in responses.

Comment: This API allows for CRUD (Create, Read, Update, and Delete) operations on the appointments table in the database. It also includes relationships with other tables and uses the include notation to include related data in responses. This API is a crucial component of the system and enables manipulation of appointment data through HTTP requests.